requirements.txt
python-telegram-bot
Pillow
selenium
